Split getMediaFile and getChildrenOf etc.
 - Leave only Get methods in MediaFileService
 - Methods that do writes are moved to WritableMediaFileService and made package-private
 - Heavy use of createOrUpdates is problematic. They have been replaced with other methods.
