+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/ShareManagementController.java
@@ -109,7 +109,7 @@ public class ShareManagementController {
             if (indexes.length == 0) {
                 return Arrays.asList(album);
             }
-            List<MediaFile> children = mediaFileService.getChildrenOf(album, true, false, true);
+            List<MediaFile> children = mediaFileService.getChildrenOf(album, true, false);
             for (int index : indexes) {
                 result.add(children.get(index));
             }
