+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/controller/SetMusicFileInfoControllerTest.java
@@ -26,9 +26,14 @@ import java.util.concurrent.ExecutionException;
 
 import com.tesshu.jpsonic.dao.AlbumDao;
 import com.tesshu.jpsonic.dao.MediaFileDao;
+import com.tesshu.jpsonic.domain.JapaneseReadingUtils;
+import com.tesshu.jpsonic.domain.JpsonicComparators;
 import com.tesshu.jpsonic.domain.MediaFile;
 import com.tesshu.jpsonic.service.MediaFileCache;
 import com.tesshu.jpsonic.service.MediaFileService;
+import com.tesshu.jpsonic.service.SecurityService;
+import com.tesshu.jpsonic.service.SettingsService;
+import com.tesshu.jpsonic.service.metadata.MetaDataParserFactory;
 import com.tesshu.jpsonic.service.scanner.ScannerStateServiceImpl;
 import com.tesshu.jpsonic.service.scanner.WritableMediaFileService;
 import org.junit.jupiter.api.BeforeEach;
@@ -46,26 +51,27 @@ import org.springframework.web.bind.ServletRequestBindingException;
 
 class SetMusicFileInfoControllerTest {
 
+    private static final int ID = 0;
+    private static final String PATH = "pathString";
     private MediaFileDao mediaFileDao;
     private ScannerStateServiceImpl scannerStateService;
     private SetMusicFileInfoController controller;
-    private int id = 0;
-    private String path = "pathString";
     private MockMvc mockMvc;
 
     @BeforeEach
     public void setup() throws ExecutionException {
         MediaFileService mediaFileService = mock(MediaFileService.class);
         MediaFile album = new MediaFile();
-        album.setId(id);
-        album.setPathString(path);
-        ;
-        Mockito.when(mediaFileService.getMediaFileStrict(id)).thenReturn(album);
+        album.setId(ID);
+        album.setPathString(PATH);
+        Mockito.when(mediaFileService.getMediaFileStrict(ID)).thenReturn(album);
 
         mediaFileDao = mock(MediaFileDao.class);
         scannerStateService = mock(ScannerStateServiceImpl.class);
         WritableMediaFileService writableMediaFileService = new WritableMediaFileService(mediaFileDao,
-                scannerStateService, mock(MediaFileService.class), mock(AlbumDao.class), mock(MediaFileCache.class));
+                scannerStateService, mock(MediaFileService.class), mock(AlbumDao.class), mock(MediaFileCache.class),
+                mock(MetaDataParserFactory.class), mock(SettingsService.class), mock(SecurityService.class),
+                mock(JapaneseReadingUtils.class), mock(JpsonicComparators.class));
         controller = new SetMusicFileInfoController(mediaFileService, writableMediaFileService);
         mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
     }
@@ -75,9 +81,9 @@ class SetMusicFileInfoControllerTest {
     void testPost() throws Exception {
         MvcResult result = mockMvc
                 .perform(MockMvcRequestBuilders.post("/setMusicFileInfo.view").param(Attributes.Request.ID.value(),
-                        Integer.toString(id)))
+                        Integer.toString(ID)))
                 .andExpect(MockMvcResultMatchers.status().isFound())
-                .andExpect(MockMvcResultMatchers.redirectedUrl(ViewName.MAIN.value() + "?id=" + id))
+                .andExpect(MockMvcResultMatchers.redirectedUrl(ViewName.MAIN.value() + "?id=" + ID))
                 .andExpect(MockMvcResultMatchers.status().is3xxRedirection()).andReturn();
         assertNotNull(result);
     }
@@ -88,26 +94,26 @@ class SetMusicFileInfoControllerTest {
         @Test
         void testNoScanning() throws ServletRequestBindingException {
             MockHttpServletRequest request = new MockHttpServletRequest();
-            request.addParameter(Attributes.Request.ID.value(), Integer.toString(id));
+            request.addParameter(Attributes.Request.ID.value(), Integer.toString(ID));
             request.addParameter(Attributes.Request.ACTION.value(), "comment");
             String comment = "Comment Test.";
             request.addParameter(Attributes.Request.COMMENT.value(), comment);
 
             controller.post(request);
-            Mockito.verify(mediaFileDao, Mockito.times(1)).updateComment(path, comment);
+            Mockito.verify(mediaFileDao, Mockito.times(1)).updateComment(PATH, comment);
         }
 
         @Test
         void testScanning() throws ServletRequestBindingException {
             MockHttpServletRequest request = new MockHttpServletRequest();
-            request.addParameter(Attributes.Request.ID.value(), Integer.toString(id));
+            request.addParameter(Attributes.Request.ID.value(), Integer.toString(ID));
             request.addParameter(Attributes.Request.ACTION.value(), "comment");
             String comment = "Comment Test.";
             request.addParameter(Attributes.Request.COMMENT.value(), comment);
 
             Mockito.when(scannerStateService.isScanning()).thenReturn(true);
             controller.post(request);
-            Mockito.verify(mediaFileDao, Mockito.times(1)).updateComment(path, comment);
+            Mockito.verify(mediaFileDao, Mockito.times(1)).updateComment(PATH, comment);
         }
     }
 
@@ -117,21 +123,21 @@ class SetMusicFileInfoControllerTest {
         @Test
         void testNoScanning() throws ServletRequestBindingException {
             MockHttpServletRequest request = new MockHttpServletRequest();
-            request.addParameter(Attributes.Request.ID.value(), Integer.toString(id));
+            request.addParameter(Attributes.Request.ID.value(), Integer.toString(ID));
             request.addParameter(Attributes.Request.ACTION.value(), "resetLastScanned");
             controller.post(request);
-            Mockito.verify(mediaFileDao, Mockito.times(1)).resetLastScanned(id);
+            Mockito.verify(mediaFileDao, Mockito.times(1)).resetLastScanned(ID);
         }
 
         @Test
         void testScanning() throws ServletRequestBindingException {
             MockHttpServletRequest request = new MockHttpServletRequest();
-            request.addParameter(Attributes.Request.ID.value(), Integer.toString(id));
+            request.addParameter(Attributes.Request.ID.value(), Integer.toString(ID));
             request.addParameter(Attributes.Request.ACTION.value(), "resetLastScanned");
 
             Mockito.when(scannerStateService.isScanning()).thenReturn(true);
             controller.post(request);
-            Mockito.verify(mediaFileDao, Mockito.never()).resetLastScanned(id);
+            Mockito.verify(mediaFileDao, Mockito.never()).resetLastScanned(ID);
         }
     }
 }
